@@ -1,23 +1,17 @@
-import logo from './logo.svg';
-import './App.css';
+import Header from './components/header/Header';
+import SearchBar from './components/search/SearchBar';
+import './app.css';
+import ForecastHeader from './components/weather/ForecastHeader';
+import Forecast from './components/weather/Forecast';
 
 function App() {
   return (
-    <div className="App">
-      <header className="App-header">
-        <img src={logo} className="App-logo" alt="logo" />
-        <p>
-          Edit <code>src/App.js</code> and save to reload.
-        </p>
-        <a
-          className="App-link"
-          href="https://reactjs.org"
-          target="_blank"
-          rel="noopener noreferrer"
-        >
-          Learn React
-        </a>
-      </header>
+    <div className="container">
+      <Header />
+      <SearchBar />
+      <hr className="hr" />
+      <ForecastHeader />
+      <Forecast />
     </div>
   );
 }
